requirements.txt
slack_bolt==1.21.2
slack_sdk==3.33.4
python-dotenv==1.0.1
schedule>=1.1.0
